Rename _maker_utils to _datamodels
